Backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
torch==2.0.1
transformers==4.30.2
Pillow==10.0.0
opencv-python==4.8.0.76